requirements.txt
tiktoken==0.7.0
pyyaml==6.0.1